perf: Make each MoveThreadRunner generate it own moves
Instead of MultiThreadedConstructionHeuristicDecider and
MultiThreadedLocalSearchDecider filling up the operationQueue
with MoveEvaluationOperation, each MoveThreadRunner will generate
it own move instead. This allows every MoveThreadRunner to evaluate
moves without waiting on the Decider. As a result, the operationQueue
will only contain ApplyStepOperation, DestroyOperation and
SetupOperation.

The MoveThreadRunner threads now will only block when:

- Waiting for the Decider to setup the step (new)
- Waiting for all MoveThreadRunner to apply a step
- Waiting for space in the result buffer
- When the iterator is exhausted

Currently, the MoveThreadRunner share the same iterator, but the code
is structured in a way so it can easily be adapted so each
MoveThreadRunner has it own independent iterator. Currently, sharing
the iterator does not have significant impact; although that might
be because we don't wait for the decider to catch up (and thus
sacrifice REPRODUCABILITY) in this commit. The next commit will add
a Semaphore to SharedNeverEndingMoveGenerator to prevent it from
generating moves until the decider catches up to the next multiple of
selectedMoveBufferSize. Using independent iterators, each with their
own independent random would remove the need of this Semaphore to
preserve REPRODUCABILITY (and thus allow the MoveThreadRunner to keep
evaluating moves, bounded only by the selectedMoveBufferSize).

OrderByMoveIndexBlockingQueue becomes the main syncing method between
the MoveThreadRunners and the Decider:

- It blocks the decider on startNextStep until all MoveThreadRunners are
  ready
- It blocks the MoveThreadRunners when all iterators are exhausted or
  when the Decider has finished a step (whichever comes first)
- It also block the MoveThreadRunner when the slot it is writing to
  has not been read by the Decider (and the Decider is correspondingly
  also blocked until a MoveThreadRunner put a result into that slot).

It employs two cyclic barriers (for moveThreadCount + 1 parties). First,
a start barrier, to prevent the move threads from corrupting the internals
that need to be fully setup before they put in new results. Second, a
end barrier, to prevent the move threads from infinitely populating
the buffer with no-op evaluations when all the iterators are exhausted.

The MoveThreadRunner generate their moves by sharing an AtomicInteger
that tracks the current move index. They atomically get and increment
that integer, and use it to select an iterator from an
AtomicReferenceArray. Move index 0 always get iterator 0, move index 1
always get iterator 1, and so on. In general, the move with index x is
generated from the iterator at index x mod moveThreadCount.
There are exactly moveThreadCount iterators, so each move thread can
concurrently generate their own moves without blocking if they use
independent iterators. In the case they use a shared iterator (i.e.
the one currently implemented), they'll only use it one at a time.
The MoveThreadRunner takes operations from the operationQueue if it is
not empty; otherwise it will create it own MoveEvaluationOperation.