fix: Remove possible deadlock from SharedNeverEndingMoveGenerator Semaphore
It is possible (although I never observed it) that SharedNeverEndingMoveGenerator
use up all its permits and prevent some MoveThreads from progressing to
resultQueue (where it need to join with the decider on a cyclic barrier).

To fix this possible deadlock, the decider will block the move threads,
drain all remaining permits (recording the number it acquired), tell the
SharedNeverEndingMoveGenerator to only return no-op's, and then release
selectedMoveBufferSize to unblock any threads waiting on the semaphore.