fix: Make BusyWaitCyclicBarrier wait for all parties to exit loop before returning
If BusyWaitCyclicBarrier did not wait for all parties to exit loop
before returning, there is a chance one party will still be in the
loop when await is called again in the future, corrupting the state.
To fix this, we use a second AtomicInteger to track how many parties
exited the loop. It will not corrupt the state if a party is still
inside the second loop when await is called again, since it will
only increment partiesWaiting count (and partiesFinished remains 0
until all parties awaited again).