chore: Use semaphore and condition instead of reseeding random
Added a new method, acquireUntil(AtomicBoolean), to BusyWaitSemaphore.
It behaves like acquire, but if there are no available permits and
the condition is set, it exits the loop without acquiring a permit.
It returns true if a permit was acquired, false otherwise.

This new method gives us a new way to guarantee reproducability:

1. At the start of each step, make the condition false and give
   each iterator moveBufferSize / iteratorCount permits
2. When the solver consumes a move evaluation, give the iterator
   that generated that move one more permit
3. When the solver decides on a move, set the condition to true,
   and wait for the move threads to reach the step end barrier.

   The iterator will be consider exhausted if one of the following
   conditions is true:

   - it has no more permits available
   - it has no more elements

   When all iterators are exhausted, the final exhausted iterator
   will signal all move threads to block on the next
   reportIteratorExhausted call. The number of next calls for each
   iterator is fixed: min(iteratorPermitCount, iteratorSize). As
   such, each iterator random will be in the same state at the end
   of each step.

This is the same strategy employed by the original algorithm, but
modified so the MoveThreadRunner can generate the moves.

Reseeding random more than likely offers better performance, but it
might affect fairness. Whereas this algorithm will perform slightly
slower, but will not affect fairness.